\name{rmarkdownJa-package}
\alias{rmarkdownJa-package}
\alias{rmarkdownJa}
\docType{package}
\title{
\packageTitle{rmarkdownJa}
}
\description{
\packageDescription{rmarkdownJa}
}
\references{
Source files at GitHub repository : \url{https://github.com/naqtn/rmarkdownJa}
}
\author{
\packageAuthor{rmarkdownJa}

Maintainer: \packageMaintainer{rmarkdownJa}
}
